changed 'small' back to Float32 to prevent domain error for run_SHMIP.jl (unknown reason)

+++ src/helpers.jl
@@ -2,7 +2,7 @@ using Base: Float64, Int64
 using LinearAlgebra, Parameters, Statistics, PyPlot
 
 # misc fudge factors to avoid dividing by zero
-small = eps(Float64)
+small = eps(Float32)
 
 """
 Turn all negative numbers into 0.0
